requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1;python_version<"3.9"
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
dj-database-url==2.3.0
Django==5.1.4
django-allauth==65.3.0
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
packaging==24.2
pillow==11.0.0
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.10.1
requests==2.32.3
sqlparse==0.5.3
stripe==11.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
whitenoise==6.8.2
